Filter class added.
