+++ src/main/Filter.java
@@ -1,0 +1,17 @@
+package main;
+
+public class Filter {
+	boolean groundsIn;
+	
+	public Filter () {
+		groundsIn = false;
+	}
+	
+	public boolean getGroundsIn() {
+		return groundsIn;
+	}
+	
+	public void setGroundsIn(boolean groundsIn) {
+		this.groundsIn = groundsIn; 
+	}
+}
